+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 * Added RegExp.execAll
 * Function.series forces asynchronicity by default (by using setImmediate),
   but if wanted this can be disabled by passing a boolean false as first arg
+* Added shim for ES6's Array#fill(value, start, end)
 
 ## 0.1.2 (2014-08-28)